Fix more warnings

+++ CesiumUtility/include/CesiumUtility/ExtensibleObject.h
@@ -128,17 +128,19 @@ struct CESIUMUTILITY_API ExtensibleObject {
     int64_t accum = 0;
     accum += sizeof(ExtensibleObject);
 
-    accum += this->extras.size() * (sizeof(std::string) + sizeof(JsonValue));
+    accum += int64_t(
+        this->extras.size() * (sizeof(std::string) + sizeof(JsonValue)));
     for (const auto& [k, v] : this->extras) {
-      accum += k.capacity() * sizeof(char) - sizeof(std::string);
-      accum += v.getSizeBytes() - static_cast<int64_t>(sizeof(JsonValue));
+      accum += int64_t(k.capacity() * sizeof(char) - sizeof(std::string));
+      accum += v.getSizeBytes() - int64_t(sizeof(JsonValue));
     }
 
-    accum += this->unknownProperties.size() *
-             (sizeof(std::string) + sizeof(JsonValue));
+    accum += int64_t(
+        this->unknownProperties.size() *
+        (sizeof(std::string) + sizeof(JsonValue)));
     for (const auto& [k, v] : this->unknownProperties) {
-      accum += k.capacity() * sizeof(char) - sizeof(std::string);
-      accum += v.getSizeBytes() - static_cast<int64_t>(sizeof(JsonValue));
+      accum += int64_t(k.capacity() * sizeof(char) - sizeof(std::string));
+      accum += v.getSizeBytes() - int64_t(sizeof(JsonValue));
     }
 
     return accum;
